worker-feedback/requirements.txt
functions-framework==3.*
vertexai
google-cloud-firestore
google-cloud-tasks
requests
certifi==2024.7.4 